Add Metadata Download Method
This adds a metadata file method to the cart. The point here is to
add supplemental metadata to a cart based on the usage of the cart.
This allows for carts to contain additional metadata from other
Pacifica services that are available to the cart.

A method (called raw) is to use the metadata contained in the
request as the document and save it with the cart.

A method (called requests) is to send a url to the cart that gets
pulled and saved by the cart upon creation. This could be a
security issue depending on the deployment, we'll address that in
a minute.

A method (called metadata) is to send a transaction ID to the cart
and based on system configuration it will query the metadata
endpoint to grab supplemental information about the transaction.
Then save that data in the cart.

Additionally there must be a configuration option to white list
certain methods as some maybe security risks for some instances.

Signed-off-by: David Brown <dmlb2000@gmail.com>